+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - 定义实体的操作服务。
   - com.dwarfeng.scg.stack.service.NodeVariableOperateService。
+  - com.dwarfeng.scg.stack.service.CommonVariableOperateService。
 
 - 定义实体及其维护服务，并通过单元测试。
   - com.dwarfeng.scg.stack.bean.entity.NodeVariable。